+++ src/models/index.ts
@@ -1,6 +1,6 @@
 export { BookmarkModel } from "./BookmarkModel";
 export { ContentModel } from "./ContentModel";
-export { SummaryModel } from "./SummaryModel";
 export { NarrationModel } from "./NarrationModel";
 export { AudioFileModel } from "./AudioFileModel";
 export { PlaylistModel } from "./PlaylistModel";
+export { MergedAudioFileModel } from "./MergedAudioFileModel";